Finding changes in blanknodes
