+++ src/main/java/com/github/checkit/model/ChangeSubjectType.java
@@ -3,5 +3,6 @@ package com.github.checkit.model;
 public enum ChangeSubjectType {
     VOCABULARY,
     TERM,
+    BLANK_NODE,
     UNKNOWN
 }
