+++ src/main/java/com/github/checkit/service/VocabularyContextService.java
@@ -9,7 +9,6 @@ import org.apache.jena.query.ParameterizedSparqlString;
 import org.apache.jena.query.Query;
 import org.apache.jena.query.QueryExecution;
 import org.apache.jena.rdf.model.Model;
-import org.apache.jena.rdf.model.Resource;
 import org.springframework.stereotype.Service;
 
 @Service
